Gen newtype lense if it has a Newtype instance
In a previous commit, we also check that the
newtype's data constructor is
exported before we generate the lens. With this commit, as long as the original source file compiles,
so will the generated lens file.

The only case where the lens file will not compile is if the Newtype reference
is imported qualified:
```
import Data.Newtype as N
import Data.Newtype (class Newtype) as N
import Data.Newtype hiding (over) as N
```
or similarly for `import Custom.Prelude as N` if the custom
prelude module re-exports `Data.Newtype (class Newtype)`